requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
dataclass-csv==1.4.0
Faker==17.0.0
idna==3.4
load-dotenv==0.1.0
numpy==1.24.2
pydash==6.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
